+++ entity/src/epilot/models/shared/getrelationsrespwithpagination.py
@@ -11,7 +11,6 @@ from typing import Any, Optional
 
 @dataclasses.dataclass
 class GetRelationsRespWithPagination:
-    r"""Success"""
     hits: Optional[float] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('hits'), 'exclude': lambda f: f is None }})
     relations: Optional[list[Any]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('relations'), 'exclude': lambda f: f is None }})