+++ entity/src/epilot/models/shared/relationitem.py
@@ -11,7 +11,6 @@ from typing import Optional
 
 @dataclasses.dataclass
 class RelationItem:
-    r"""Success"""
     attribute: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('attribute') }})
     entity_id: str = dataclasses.field(metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('entity_id') }})
     tags: Optional[list[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('_tags'), 'exclude': lambda f: f is None }})